src/modules/70-pipeline/components/ArtifactsSelection/ArtifactRepository/ArtifactLastSteps/GoogleArtifactRegistry/__tests__/GoogleArtifactRegistry.test.tsx
/*
 * Copyright 2022 Harness Inc. All rights reserved.
 * Use of this source code is governed by the PolyForm Shield 1.0.0 license
 * that can be found in the licenses directory at the root of this repository, also available at
 * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
 */

import React from 'react'
import {
  act,
  findByText,
  fireEvent,
  queryByAttribute,
  render,
  waitFor,
  getByText as getElementByText
} from '@testing-library/react'
import userEvent from '@testing-library/user-event'
import { AllowedTypesWithRunTime, MultiTypeInputType, RUNTIME_INPUT_VALUE, StepProps } from '@harness/uicore'
import type * as cdng from 'services/cd-ng'

import { TestWrapper } from '@common/utils/testUtils'
import {
  ArtifactType,
  TagTypes,
  GoogleArtifactRegistryInitialValuesType,
  GoogleArtifactRegistryProps
} from '@pipeline/components/ArtifactsSelection/ArtifactInterface'
import { ModalViewFor } from '@pipeline/components/ArtifactsSelection/ArtifactHelper'
import { GoogleArtifactRegistry } from '../GoogleArtifactRegistry'

const commonInitialValues: GoogleArtifactRegistryInitialValuesType = {
  identifier: '',
  versionType: TagTypes.Value,
  spec: {
    connectorRef: '',
    package: '',
    repositoryType: 'docker',
    project: '',
    region: { label: '', value: '' } as any,
    repositoryName: '',
    version: ''
  }
}

const repoListData = {
  status: 'SUCCESS',
  data: {
    garRepositoryDTOList: [
      {
        repository: 'testRepo',
        format: 'DOCKER'
      },
      {
        repository: 'testRepo 2',
        format: 'DOCKER'
      }
    ]
  }
}

const packageListData = {
  status: 'SUCCESS',
  data: {
    garPackageDTOList: [
      {
        packageName: 'testProject',
        createTime: '2022-09-14T03:10:05.836970Z',
        updateTime: '2022-09-14T03:10:05.836970Z'
      },
      {
        packageName: 'testProject 2',
        createTime: '2022-09-14T03:10:05.836970Z',
        updateTime: '2022-09-14T03:10:05.836970Z'
      }
    ]
  }
}

const buildData = {
  status: 'SUCCESS',
  data: {
    buildDetailsList: [
      {
        version: 'v3.0'
      },
      {
        version: 'v1.0'
      },
      {
        version: 'latest'
      }
    ]
  },
  metaData: null,
  correlationId: '441c6388-e3df-44cd-86f8-ccc6f1a4558b'
}

const regionData = {
  status: 'SUCCESS',
  data: [
    {
      name: 'us-east',
      value: 'us-east'
    },
    {
      name: 'us-east1',
      value: 'us-east1'
    }
  ],
  metaData: null,
  correlationId: '441c6388-e3df-44cd-86f8-ccc6f1a4558b'
}
const fetchRepos = jest.fn().mockReturnValue(repoListData)
const fetchPackages = jest.fn().mockReturnValue(packageListData)
const fetchBuilds = jest.fn().mockReturnValue(buildData)

jest.mock('services/cd-ng', () => ({
  useGetRepositoriesForGoogleArtifactRegistry: jest.fn().mockImplementation(() => {
    return { data: repoListData, refetch: fetchRepos, error: null, loading: false }
  }),
  useGetPackagesForGoogleArtifactRegistry: jest.fn().mockImplementation(() => {
    return { data: packageListData, refetch: fetchPackages, error: null, loading: false }
  }),
  useGetBuildDetailsForGoogleArtifactRegistry: jest.fn().mockImplementation(() => {
    return { data: buildData, refetch: fetchBuilds, error: null, loading: false }
  }),
  useGetRegionsForGoogleArtifactRegistry: jest.fn().mockImplementation(() => {
    return { data: regionData }
  }),
  useGetServiceV2: jest.fn().mockImplementation(() => ({ loading: false, data: {}, refetch: jest.fn() })),
  useGetLastSuccessfulBuildForGoogleArtifactRegistry: jest.fn().mockImplementation(() => {
    return {
      data: {
        metadata: {
          SHA: 'test',
          SHAV2: 'test2'
        }
      },
      refetch: jest.fn(),
      error: null,
      loading: false
    }
  })
}))

const onSubmit = jest.fn()
export const props: Omit<StepProps<cdng.ConnectorConfigDTO> & GoogleArtifactRegistryProps, 'initialValues'> = {
  key: 'key',
  name: 'Artifact details',
  expressions: [],
  allowableTypes: [
    MultiTypeInputType.FIXED,
    MultiTypeInputType.RUNTIME,
    MultiTypeInputType.EXPRESSION
  ] as AllowedTypesWithRunTime[],
  context: 1,
  handleSubmit: onSubmit,
  artifactIdentifiers: [],
  selectedArtifact: 'GoogleArtifactRegistry' as ArtifactType,
  prevStepData: {
    connectorId: {
      value: 'testConnector'
    }
  }
}

const doConfigureOptionsTesting = async (cogModal: HTMLElement, fieldElement: HTMLInputElement) => {
  // Type regex and submit
  // check if field has desired value
  await waitFor(() => expect(getElementByText(cogModal, 'common.configureOptions.regex')).toBeInTheDocument())
  const regexRadio = getElementByText(cogModal, 'common.configureOptions.regex')
  await userEvent.click(regexRadio)
  const regexTextArea = queryByAttribute('name', cogModal, 'regExValues')
  act(() => {
    fireEvent.change(regexTextArea!, { target: { value: '<+input>.includes(/test/)' } })
  })
  const cogSubmit = getElementByText(cogModal, 'submit')
  await userEvent.click(cogSubmit)
  await waitFor(() => expect(fieldElement.value).toBe('<+input>.regex(<+input>.includes(/test/))'))
}

describe('GoogleArtifactRegistry tests', () => {
  beforeEach(() => {
    fetchBuilds.mockReset()
    onSubmit.mockReset()
    fetchRepos.mockReset()
    fetchPackages.mockReset()
  })
  test(`renders fine for the NEW artifact`, () => {
    const { container } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={commonInitialValues} {...props} />
      </TestWrapper>
    )
    expect(container).toMatchSnapshot()
  })

  test(`renders fine for the versionType as value`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      spec: {
        ...commonInitialValues.spec,
        project: 'testProject',
        package: 'testPackage',
        region: { label: 'us-east1', value: 'us-east1' } as any,
        repositoryName: 'testRepo',
        version: 'xyz.zip'
      }
    }
    const { container, getByText } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )

    const queryByNameAttribute = (name: string): HTMLElement | null => queryByAttribute('name', container, name)
    expect(queryByNameAttribute('spec.project')).not.toBeNull()
    expect(queryByNameAttribute('spec.region')).not.toBeNull()
    expect(queryByNameAttribute('spec.repositoryName')).not.toBeNull()
    expect(queryByNameAttribute('spec.version')).not.toBeNull()
    expect(container).toMatchSnapshot()

    const submitBtn = getByText('submit')
    fireEvent.click(submitBtn)
    await waitFor(() => {
      expect(props.handleSubmit).toBeCalled()
      expect(props.handleSubmit).toHaveBeenCalledWith({
        spec: {
          connectorRef: 'testConnector',
          project: 'testProject',
          package: 'testPackage',

          region: 'us-east1',
          repositoryName: 'testRepo',
          repositoryType: 'docker',
          version: 'xyz.zip'
        }
      })
    })
  })

  test(`renders fine for sidecar artifact`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      identifier: 'initial_id',
      spec: {
        ...commonInitialValues.spec,
        project: 'testProject',
        package: 'testPackage',

        region: { label: 'us-east1', value: 'us-east1' } as any,
        repositoryName: 'testRepo',
        version: 'xyz.zip'
      }
    }
    const { container, getByText } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} context={ModalViewFor.SIDECAR} />
      </TestWrapper>
    )

    const queryByNameAttribute = (name: string): HTMLElement | null => queryByAttribute('name', container, name)

    const identifierField = queryByNameAttribute('identifier') as HTMLInputElement
    expect(identifierField.value).toBe('initial_id')
    // change value of identifier to empty
    act(() => {
      fireEvent.change(queryByNameAttribute('identifier')!, { target: { value: '' } })
    })

    const submitBtn = getByText('submit')
    fireEvent.click(submitBtn)

    const identifierRequiredErr = await findByText(container, 'common.validation.fieldIsRequired')
    expect(identifierRequiredErr).toBeDefined()

    // change value of identifier
    act(() => {
      fireEvent.change(queryByNameAttribute('identifier')!, { target: { value: 'test_id' } })
    })

    fireEvent.click(submitBtn)
    await waitFor(() => {
      expect(props.handleSubmit).toBeCalled()
      expect(props.handleSubmit).toHaveBeenCalledWith({
        identifier: 'test_id',
        spec: {
          connectorRef: 'testConnector',
          project: 'testProject',
          package: 'testPackage',

          region: 'us-east1',
          repositoryName: 'testRepo',
          repositoryType: 'docker',
          version: 'xyz.zip'
        }
      })
    })
  })

  test(`renders fine for the versionType as regex`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      versionType: TagTypes.Regex,
      spec: {
        ...commonInitialValues.spec,
        project: 'testProject',
        package: 'testPackage',

        region: { label: 'us-east1', value: 'us-east1' } as any,
        repositoryName: 'testRepo',
        versionRegex: '*.zip'
      }
    }
    const { container, getByText } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )

    const queryByNameAttribute = (name: string): HTMLElement | null => queryByAttribute('name', container, name)
    expect(queryByNameAttribute('spec.project')).not.toBeNull()
    expect(queryByNameAttribute('spec.region')).not.toBeNull()
    expect(queryByNameAttribute('spec.repositoryName')).not.toBeNull()
    expect(queryByNameAttribute('spec.versionRegex')).not.toBeNull()
    expect(container).toMatchSnapshot()

    const submitBtn = getByText('submit')
    fireEvent.click(submitBtn)
    await waitFor(() => {
      expect(props.handleSubmit).toBeCalled()
      expect(props.handleSubmit).toHaveBeenCalledWith({
        spec: {
          connectorRef: 'testConnector',
          package: 'testPackage',

          project: 'testProject',
          region: 'us-east1',
          repositoryName: 'testRepo',
          repositoryType: 'docker',
          versionRegex: '*.zip'
        }
      })
    })
  })

  test(`configure values should work fine when all values are runtime input and versionType as value`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      spec: {
        ...commonInitialValues.spec,
        project: RUNTIME_INPUT_VALUE,
        package: RUNTIME_INPUT_VALUE,

        region: RUNTIME_INPUT_VALUE,
        repositoryName: RUNTIME_INPUT_VALUE,
        version: RUNTIME_INPUT_VALUE
      }
    }
    const { container } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )

    const queryByNameAttribute = (name: string): HTMLElement | null => queryByAttribute('name', container, name)
    const versionInput = queryByNameAttribute('spec.version') as HTMLInputElement
    expect(versionInput).not.toBeNull()

    const modals = document.getElementsByClassName('bp3-dialog')
    expect(modals.length).toBe(0)

    // Configure options testing for bucketName and filePath fields
    const cogVersion = document.getElementById('configureOptions_version')
    await userEvent.click(cogVersion!)
    await waitFor(() => expect(modals.length).toBe(1))
    const versionCOGModal = modals[0] as HTMLElement
    await doConfigureOptionsTesting(versionCOGModal, versionInput)

    const submitBtn = getElementByText(container, 'submit')
    fireEvent.click(submitBtn)
    await waitFor(() => {
      expect(props.handleSubmit).toBeCalled()
      expect(props.handleSubmit).toHaveBeenCalledWith({
        spec: {
          connectorRef: 'testConnector',
          project: '<+input>',
          package: '<+input>',
          region: '<+input>',
          repositoryName: '<+input>',
          repositoryType: 'docker',
          version: '<+input>.regex(<+input>.includes(/test/))'
        }
      })
    })
  })

  test(`configure values should work fine when all values are runtime input and versionType as regex`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      versionType: TagTypes.Regex,
      spec: {
        ...commonInitialValues.spec,
        project: RUNTIME_INPUT_VALUE,
        package: RUNTIME_INPUT_VALUE,
        region: RUNTIME_INPUT_VALUE,
        repositoryName: RUNTIME_INPUT_VALUE,
        versionRegex: RUNTIME_INPUT_VALUE
      }
    }
    const { container } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )

    const queryByNameAttribute = (name: string): HTMLElement | null => queryByAttribute('name', container, name)
    const projectInput = queryByNameAttribute('spec.project') as HTMLInputElement
    const regionInput = queryByNameAttribute('spec.region') as HTMLInputElement
    const packageInput = queryByNameAttribute('spec.package') as HTMLInputElement
    const repositoryNameInput = queryByNameAttribute('spec.repositoryName') as HTMLInputElement
    const versionRegexInput = queryByNameAttribute('spec.versionRegex') as HTMLInputElement
    expect(projectInput).not.toBeNull()
    expect(regionInput).not.toBeNull()
    expect(repositoryNameInput).not.toBeNull()
    expect(versionRegexInput).not.toBeNull()

    const modals = document.getElementsByClassName('bp3-dialog')
    expect(modals.length).toBe(0)

    // Configure options testing for bucketName and filePath fields
    const cogProject = document.getElementById('configureOptions_project')
    await userEvent.click(cogProject!)
    await waitFor(() => expect(modals.length).toBe(1))
    const projectCOGModal = modals[0] as HTMLElement
    await doConfigureOptionsTesting(projectCOGModal, projectInput)

    const cogRegion = document.getElementById('configureOptions_region')
    await userEvent.click(cogRegion!)
    await waitFor(() => expect(modals.length).toBe(1))
    const regionCOGModal = modals[0] as HTMLElement
    await doConfigureOptionsTesting(regionCOGModal, regionInput)

    const cogPackage = document.getElementById('configureOptions_package')
    await userEvent.click(cogPackage!)
    await waitFor(() => expect(modals.length).toBe(1))
    const packageCOGModal = modals[0] as HTMLElement
    await doConfigureOptionsTesting(packageCOGModal, packageInput)

    const cogRepositoryName = document.getElementById('configureOptions_repositoryName')
    await userEvent.click(cogRepositoryName!)
    await waitFor(() => expect(modals.length).toBe(1))
    const repositoryNameCOGModal = modals[0] as HTMLElement
    await doConfigureOptionsTesting(repositoryNameCOGModal, repositoryNameInput)

    const cogVersionRegex = document.getElementById('configureOptions_versionRegex')
    await userEvent.click(cogVersionRegex!)
    await waitFor(() => expect(modals.length).toBe(2))
    const versionRegexCOGModal = modals[1] as HTMLElement
    await doConfigureOptionsTesting(versionRegexCOGModal, versionRegexInput)

    const submitBtn = getElementByText(container, 'submit')
    fireEvent.click(submitBtn)
    await waitFor(() => {
      expect(props.handleSubmit).toBeCalled()
      expect(props.handleSubmit).toHaveBeenCalledWith({
        spec: {
          connectorRef: 'testConnector',
          project: '<+input>.regex(<+input>.includes(/test/))',
          package: '<+input>.regex(<+input>.includes(/test/))',
          region: '<+input>.regex(<+input>.includes(/test/))',
          repositoryName: '<+input>.regex(<+input>.includes(/test/))',
          repositoryType: 'docker',
          versionRegex: '<+input>.regex(<+input>.includes(/test/))'
        }
      })
    })
  })

  test(`clicking on version list should fetch builds`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      spec: {
        ...commonInitialValues.spec,
        project: 'testProject',
        package: 'testPackage',
        region: 'testRegion',
        repositoryName: 'testRepo',
        version: 'xyz'
      }
    }
    const { container } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )
    const dropdownIcons = container.querySelectorAll('[data-icon="chevron-down"]')
    const versionNameDropDownIcon = dropdownIcons[4]
    act(() => {
      fireEvent.click(versionNameDropDownIcon)
    })
    await waitFor(() => expect(fetchBuilds).toHaveBeenCalled())
  })

  test(`clicking on version list should not fetch builds if required fields are empty`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      spec: {
        ...commonInitialValues.spec,
        project: 'testProject',
        package: '',
        region: 'testRegion',
        repositoryName: '',
        version: 'xyz'
      }
    }
    const { container } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )
    const dropdownIcons = container.querySelectorAll('[data-icon="chevron-down"]')
    const versionNameDropDownIcon = dropdownIcons[3]
    act(() => {
      fireEvent.click(versionNameDropDownIcon)
    })
    await waitFor(() => expect(fetchBuilds).not.toHaveBeenCalled())
  })

  test(`clicking on repository name list should fetch repo list`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      spec: {
        ...commonInitialValues.spec,
        project: 'testProject',
        package: 'testPackage',
        region: 'testRegion',
        repositoryName: 'testRepo',
        version: 'xyz'
      }
    }
    const { container } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )
    const dropdownIcons = container.querySelectorAll('[data-icon="chevron-down"]')
    const repositoryNameDropDownIcon = dropdownIcons[2]
    act(() => {
      fireEvent.click(repositoryNameDropDownIcon)
    })
    await waitFor(() =>
      expect(fetchRepos).toHaveBeenCalledWith({
        queryParams: {
          connectorRef: 'testConnector',
          project: 'testProject',
          region: 'testRegion'
        }
      })
    )
  })

  test(`clicking on repository name list should not fetch repo list if required fields are empty`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      spec: {
        ...commonInitialValues.spec,
        project: '',
        package: '',
        region: '',
        repositoryName: 'testRepo',
        version: 'xyz'
      }
    }
    const { container } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )
    const dropdownIcons = container.querySelectorAll('[data-icon="chevron-down"]')
    const repositoryNameDropDownIcon = dropdownIcons[2]
    act(() => {
      fireEvent.click(repositoryNameDropDownIcon)
    })
    await waitFor(() => expect(fetchRepos).not.toHaveBeenCalled())
  })
  test(`clicking on package list should fetch package list`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      spec: {
        ...commonInitialValues.spec,
        project: 'testProject',
        package: 'testPackage',
        region: 'testRegion',
        repositoryName: 'testRepo',
        version: 'xyz'
      }
    }
    const { container } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )
    const dropdownIcons = container.querySelectorAll('[data-icon="chevron-down"]')
    const packageDropDownIcon = dropdownIcons[3]
    act(() => {
      fireEvent.click(packageDropDownIcon)
    })
    await waitFor(() => expect(fetchPackages).toHaveBeenCalled())
  })

  test(`clicking on pacakge list should not fetch package list if required fields are empty`, async () => {
    const initialValues = {
      type: 'GoogleArtifactRegistry',
      ...commonInitialValues,
      spec: {
        ...commonInitialValues.spec,
        project: '',
        region: '',
        repositoryName: '',
        package: 'test',
        version: 'xyz'
      }
    }
    const { container } = render(
      <TestWrapper>
        <GoogleArtifactRegistry initialValues={initialValues as any} {...props} />
      </TestWrapper>
    )
    const dropdownIcons = container.querySelectorAll('[data-icon="chevron-down"]')
    const packageDropDownIcon = dropdownIcons[3]
    act(() => {
      fireEvent.click(packageDropDownIcon)
    })
    await waitFor(() => expect(fetchPackages).not.toHaveBeenCalled())
  })
})
